Change test to use mpt directly
